refactor(views): extract coordinators into dedicated files
Move GenerateMenuCoordinator and AddRecipeCoordinator out of their view
files and rename load(from:) to loadExistingRecipe(from:) for clarity.
